moved WPTheme/PluginConfig to dedicated files themes/plugin.py
