+++ src/wordpress/__init__.py
@@ -1,5 +1,7 @@
 from .models import WPException, WPSite, WPUser
-from .config import WPConfig, WPPluginConfig, WPThemeConfig
+from .config import WPConfig
+from .themes import WPThemeConfig
+from .plugins import WPPluginConfig
 from .generator import WPGenerator
 
 
@@ -7,8 +9,8 @@ __all__ = [
     WPException,
     WPSite,
     WPUser,
+    WPGenerator,
     WPConfig,
     WPPluginConfig,
     WPThemeConfig,
-    WPGenerator
 ]
